refactor: split mock server helpers into separate module
- Create helpers_mock_server.rs for all mock server setup functions
- Move 9 mock server helper functions from helpers.rs to new module
- Move solver registry resource creation functions to new module
- Update mod.rs to include and re-export new module
- Reduce helpers.rs from ~617 to ~317 lines for better organization

Tests pass: Move 62, EVM 48, Verifier (Rust) 178, Solver (Rust) 119
